added whenCondition to damage
Reorganized files too. Also added messages when importing
